crossover2 function added
Crossover 2 takes a random city and finds it's position in the parent's
order vector, then choses one parent randomly but weighted based on
their performance and places thic city in the output vector in this
position. If the first choice parent's position is taken alread the
second choice is used, if both are taken it is placed randomly at the
end. Mutations prevent the whole population becoming identical and
introduce new samples occasionall. Works well with weightmapping -
replaces crossover in sketch.